requirements.txt
discord.py
lxml
pytz